tests/data_test_main/user_inputs/mvs_inputs/csv_elements/fixcost.csv

+++ tests/test_check_inputs.py
@@ -317,7 +317,7 @@ class TestDemandProfiles:
         file = pd.read_csv(filename, index_col=0, header=0,)
 
         assert (
-            int(file.at["evaluated_period", "simulation_settings"])
+            float(file.at["evaluated_period", "simulation_settings"])
             == len(ts.index) / 24
         )
